Commit before Rewrite of Fundamental Call of RestSharp Execute Method
All error responses were throwing HttpRestExceptions and swallowing the response that I needed to process for weirdness during the starting of Connect strategies. Getting ready to implement the execute request method properly so that I can handle ALL responses with my handlers and return the information needed to the caller.